Use UUID's for CFDataset's unique_id property
Constructing the unique_id from attributes of the dataset leads to
problems. One can have multiple datasets that end up with the same
unique_id (for example datasets that have a different spatial domain,
but everything else is the same). When generated according to the
standard methods, UUIDs are for practical purposes unique.